Add pagination

+++ article/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render,get_object_or_404,redirect
 from django.views.decorators.csrf import csrf_protect
 from django.views.generic.edit import CreateView
+from django.core.paginator import Paginator
 
 
 # Create your views here.
@@ -11,19 +12,31 @@ import markdown
 from .forms import ArticlePostForm
 # 引入User模型
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+
 
 @csrf_protect
 def article_list(request):
 
-    articles = ArticlePost.objects.all()
-    context = {'articles': articles }
+    article_list = ArticlePost.objects.all()
+    paginator = Paginator(article_list, 3)
+    page = request.GET.get('page')
+    
+    articles = paginator.get_page(page)
 
+    context = { 'articles': articles }
     return render(request, 'article/list.html', context)
+
 
 
 @csrf_protect
 def article_detail(request, id):
     article = ArticlePost.objects.get(id=id)
+
+    if request.user!= article.author:
+        article.total_views += 1
+        article.save(update_fields=['total_views'])
+
 
     # 将markdown语法渲染成html样式
     article.body = markdown.markdown(article.body,
@@ -69,9 +82,13 @@ def article_create(request):
         return render(request, 'article/create.html', context)
 
 @csrf_protect
+@login_required(login_url='/userprofile/login/')
 def article_update(request, id):
 
     article = ArticlePost.objects.get(id=id)
+
+    if request.user!= article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
 
     if request.method == "POST":
         article_post_form = ArticlePostForm(data=request.POST)
@@ -96,11 +113,16 @@ def article_update(request, id):
 
 
 @csrf_protect
+@login_required(login_url='/userprofile/login/')
 def article_safe_delete(request, id):
     if request.method == 'POST':
         article = ArticlePost.objects.get(id=id)
-        article.delete()
-        return redirect("article:article_list")
+
+        if request.user!= article.author:
+            return HttpResponse("对不起，你无权修改这篇文章。")
+        else:
+            article.delete()
+            return redirect("article:article_list")
     else:
         return HttpResponse("仅允许post请求")
